requirements.txt
# Python 3.10+ recommended
# Install: pip install -r requirements.txt

# for Kaggle
#   !pip install -r requirements.txt -q

#opencv-python>=4.8.0
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0 
timm>=0.9.12
tensorflow
numpy>=1.24.0
pandas>=2.0.0              
Pillow>=10.0.0             
scikit-learn>=1.3.0        
PyYAML>=6.0                # loads config.yaml in every script
matplotlib>=3.7.0
seaborn>=0.12.0 
tqdm>=4.65.0               # progress bars during training  
# wandb>=0.16.0            # uncomment if team decides to use Weights & Biases
gradio>=4.0.0            # when building the demo
 